+++ src/YammerShell/YammerObjects/YammerUser.cs
@@ -5,7 +5,7 @@ namespace YammerShell.YammerObjects
 {
     public class YammerUser
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string UserName { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
@@ -14,7 +14,7 @@ namespace YammerShell.YammerObjects
         public string JobTitle { get; set; }
         public string Department { get; set; }
         public string Timezone { get; set; }
-        public int NetworkId { get; set; }
+        public long NetworkId { get; set; }
         public string NetworkName { get; set; }
         public IEnumerable<string> PhoneNumbers { get; set; }
         public DateTime ActivatedAt { get; set; }
